WIP continue fixing tests for modularizing
