+++ examples/channels/index.html
@@ -33,7 +33,7 @@
 
 		require([
 			'seriously',
-			'src/effects/seriously.channels',
+			'src/effects/channels',
 			'src/effects/seriously.blend'
 		], function (Seriously) {
 			// declare our variables
