+++ examples/demo/fog.html
@@ -50,7 +50,7 @@
 	<script type="text/javascript" src="../../seriously.js"></script>
 	<script type="text/javascript" src="../../src/sources/seriously.depthmap.js"></script>
 	<script type="text/javascript" src="../../src/effects/seriously.simplex.js"></script>
-	<script type="text/javascript" src="../../src/effects/seriously.channels.js"></script>
+	<script type="text/javascript" src="../../src/effects/channels.js"></script>
 	<script type="text/javascript" src="../../src/effects/seriously.falsecolor.js"></script>
 	<script type="text/javascript" src="../../src/effects/seriously.blend.js"></script>
 	<script type="text/javascript" src="../../src/effects/seriously.linear-transfer.js"></script>
