+++ index.html
@@ -12,7 +12,7 @@
 		<script type="text/javascript" src="src/effects/seriously.blend.js"></script>
 		<script type="text/javascript" src="src/effects/seriously.brightness-contrast.js"></script>
 		<script type="text/javascript" src="src/effects/seriously.checkerboard.js"></script>
-		<script type="text/javascript" src="src/effects/seriously.chroma.js"></script>
+		<script type="text/javascript" src="src/effects/chroma.js"></script>
 		<script type="text/javascript" src="src/effects/seriously.color.js"></script>
 		<script type="text/javascript" src="src/effects/seriously.colorcomplements.js"></script>
 		<script type="text/javascript" src="src/effects/seriously.color-select.js"></script>
@@ -32,7 +32,7 @@
 		<script type="text/javascript" src="src/effects/seriously.blur.js"></script>
 		<script type="text/javascript" src="src/effects/seriously.hue-saturation.js"></script>
 		<script type="text/javascript" src="src/effects/seriously.hex.js"></script>
-		<script type="text/javascript" src="src/effects/seriously.invert.js"></script>
+		<script type="text/javascript" src="src/effects/invert.js"></script>
 		<script type="text/javascript" src="src/effects/seriously.directionblur.js"></script>
 		<script type="text/javascript" src="src/effects/seriously.nightvision.js"></script>
 		<script type="text/javascript" src="src/effects/seriously.noise.js"></script>
