+++ src/constants.js
@@ -86,6 +86,7 @@ const colorNames = {
 		'defaults',
 		'destroy',
 		'effect',
+		'draw',
 		'go',
 		'id',
 		'incompatible',
@@ -95,12 +96,29 @@ const colorNames = {
 		'isSource',
 		'isTarget',
 		'isTransform',
+		'initDaemon',
+		'addAlias',
 		'removeAlias',
 		'render',
 		'source',
 		'stop',
+		'addNode',
+		'removeNode',
 		'target',
-		'transform'
+		'transform',
+		'getNodeId',
+		'findInputNode',
+		'addSourceNode',
+		'removeSourceNode',
+		'addEffectNode',
+		'removeEffectNode',
+		'addTransformNode',
+		'removeTransformNode',
+		'addTargetNode',
+		'removeTargetNode',
+		'commonShaders',
+		'auto',
+		'attachContext'
 	],
 
 	baseVertexShader = [
