+++ test/index.js
@@ -1,6 +1,11 @@
 import Seriously from '../src/seriously';
-import '../src/transforms/transform2d';
 import '../src/sources/array';
 import '../src/sources/imagedata';
+import '../src/sources/video';
+import '../src/transforms/transform2d';
+import '../src/transforms/reformat';
+import '../src/effects/chroma';
+import '../src/effects/invert';
+import '../src/effects/channels';
 
 export default Seriously;
