+++ test/seriously.unit.js
@@ -2971,231 +2971,207 @@
 	use require for loading plugins
 	*/
 	module('Effect Plugins');
-	asyncTest('Invert', 3, function () {
-		require([
-			'seriously',
-			'effects/seriously.invert',
-			'src/sources/array'
-		], function (Seriously) {
-			var seriously,
-				effect,
-				target,
-				canvas,
-				source,
-				pixels,
-				error,
-				incompatible,
-				input = [255, 128, 100, 200],
-				expected;
+	test('Invert', 3, function () {
+		var seriously,
+			effect,
+			target,
+			canvas,
+			source,
+			pixels,
+			error,
+			incompatible,
+			input = [255, 128, 100, 200],
+			expected;
 
-			incompatible = Seriously.incompatible();
+		incompatible = Seriously.incompatible();
 
-			seriously = new Seriously();
-			source = seriously.source(input, {
-				width: 1,
-				height: 1
-			});
+		seriously = new Seriously();
+		source = seriously.source(input, {
+			width: 1,
+			height: 1
+		});
 
-			canvas = document.createElement('canvas');
-			canvas.width = canvas.height = 1;
-			target = seriously.target(canvas);
+		canvas = document.createElement('canvas');
+		canvas.width = canvas.height = 1;
+		target = seriously.target(canvas);
 
-			effect = seriously.effect('invert');
+		effect = seriously.effect('invert');
 
-			ok(effect, 'Invert effect successfully created');
+		ok(effect, 'Invert effect successfully created');
 
-			target.source = effect;
-			effect.source = source;
+		target.source = effect;
+		effect.source = source;
+
+		try {
+			pixels = target.readPixels(0, 0, 1, 1);
+		} catch (e) {
+			error = e;
+		}
+
+		expected = input.map(function (value, channel) {
+			if (channel < 3) {
+				return Math.round((255 - value) * input[3] / 255);
+			}
+
+			return value;
+		});
+
+		ok(incompatible ? error : !error, 'readPixels throws error iff incompatible');
+		ok(incompatible || pixels && compare(pixels, expected), 'Invert effect rendered accurately.');
+
+		seriously.destroy();
+		Seriously.removePlugin('invert');
+	});
+
+	test('Chroma Key', function (Seriously) {
+		var seriously,
+			effect,
+			target,
+			canvas,
+			source,
+			pixels,
+			error,
+			incompatible,
+
+			colors = [
+				[31, 57, 28], // low saturation, any balance
+				[28, 57, 33], // low saturation only when balance = 0
+				[0, 23, 0], // low maximum channel
+				[51, 230, 77], // high saturation
+				[51, 179, 77] // mid saturation (partial alpha)
+			],
+			sourcePixels = [];
+
+		colors.forEach(function (color, index) {
+			var i;
+
+			// green
+			i = index * 4;
+			sourcePixels[i + 0] = color[0];// / 255;
+			sourcePixels[i + 1] = color[1];// / 255;
+			sourcePixels[i + 2] = color[2];// / 255;
+			sourcePixels[i + 3] = 255; // alpha
+
+			// blue
+			i = index * 4 + colors.length * 4;
+			sourcePixels[i + 0] = color[0];// / 255;
+			sourcePixels[i + 1] = color[2];// / 255;
+			sourcePixels[i + 2] = color[1];// / 255;
+			sourcePixels[i + 3] = 255; // alpha
+		});
+
+		incompatible = Seriously.incompatible('chroma');
+
+		expect(incompatible ? 3 : 23);
+
+		seriously = new Seriously();
+		source = seriously.source(sourcePixels, {
+			width: colors.length,
+			height: 2
+		});
+
+		canvas = document.createElement('canvas');
+		canvas.width = 5;
+		canvas.height = 2;
+		target = seriously.target(canvas);
+
+		effect = seriously.effect('chroma');
+		ok(effect, 'Chroma effect successfully created');
+
+		effect.source = source;
+		target.source = effect;
+		effect.balance = 0;
+		effect.clipBlack = 0.1;
+		effect.clipWhite = 0.88;
+
+		// test with both green and blue screen colors
+		['rgb(51, 204, 77)', 'rgb(51, 77, 204)'].forEach(function (screen, row) {
+			var matchOffset = row * colors.length * 4,
+				noMatchOffset = (1 - row) * colors.length * 4,
+				color = row ? 'Blue' : 'Green';
+
+			effect.screen = screen;
+			target.render();
 
 			try {
-				pixels = target.readPixels(0, 0, 1, 1);
+				pixels = target.readPixels(0, 0, colors.length, 2, pixels);
 			} catch (e) {
 				error = e;
 			}
-
-			expected = input.map(function (value, channel) {
-				if (channel < 3) {
-					return Math.round((255 - value) * input[3] / 255);
-				}
-
-				return value;
-			});
-
 			ok(incompatible ? error : !error, 'readPixels throws error iff incompatible');
-			ok(incompatible || pixels && compare(pixels, expected), 'Invert effect rendered accurately.');
+			if (!incompatible) {
+				equal(pixels[matchOffset + 3], 247, color + ' - Low saturation partially keyed out');
+				equal(pixels[matchOffset + 7], 255, color + ' - Very low saturation remains opaque');
+				equal(pixels[matchOffset + 11], 255, color + ' - Low saturation and max channel remains opaque');
+				equal(pixels[matchOffset + 15], 0, color + ' - High saturation fully keyed out');
+				equal(pixels[matchOffset + 19], 11, color + ' - Mid saturation partially keyed out');
 
-			seriously.destroy();
-			Seriously.removePlugin('invert');
-
-			start();
+				equal(pixels[noMatchOffset + 3], 255, color + ' - Low saturation, different color remains opaque');
+				equal(pixels[noMatchOffset + 7], 255, color + ' - Very low saturation, different color remains opaque');
+				equal(pixels[noMatchOffset + 11], 255, color + ' - Low saturation and max channel, different color remains opaque');
+				equal(pixels[noMatchOffset + 15], 255, color + ' - High saturation, different color remains opaque');
+				equal(pixels[noMatchOffset + 19], 255, color + ' - Mid saturation, different color remains opaque');
+			}
 		});
+
+
+		seriously.destroy();
+		Seriously.removePlugin('chroma');
 	});
 
-	asyncTest('Chroma Key', function () {
-		require([
-			'seriously',
-			'effects/seriously.chroma',
-			'src/sources/array'
-		], function (Seriously) {
-			var seriously,
-				effect,
-				target,
-				canvas,
-				source,
-				pixels,
-				error,
-				incompatible,
+	test('Channels', 5, function (Seriously) {
+		var seriously,
+			effect,
+			target,
+			canvas,
+			pixels,
+			error,
+			incompatible,
+			array = [
+				255, 255, 255, 255,
+				255, 0, 0, 255,
+				0, 255, 0, 128,
+				0, 0, 255, 0
+			],
+			arraySource,
+			img,
+			imgSource,
+			expected;
 
-				colors = [
-					[31, 57, 28], // low saturation, any balance
-					[28, 57, 33], // low saturation only when balance = 0
-					[0, 23, 0], // low maximum channel
-					[51, 230, 77], // high saturation
-					[51, 179, 77] // mid saturation (partial alpha)
-				],
-				sourcePixels = [];
+		incompatible = Seriously.incompatible();
 
-			colors.forEach(function (color, index) {
-				var i;
-
-				// green
-				i = index * 4;
-				sourcePixels[i + 0] = color[0];// / 255;
-				sourcePixels[i + 1] = color[1];// / 255;
-				sourcePixels[i + 2] = color[2];// / 255;
-				sourcePixels[i + 3] = 255; // alpha
-
-				// blue
-				i = index * 4 + colors.length * 4;
-				sourcePixels[i + 0] = color[0];// / 255;
-				sourcePixels[i + 1] = color[2];// / 255;
-				sourcePixels[i + 2] = color[1];// / 255;
-				sourcePixels[i + 3] = 255; // alpha
-			});
-
-			incompatible = Seriously.incompatible('chroma');
-
-			expect(incompatible ? 3 : 23);
-
-			seriously = new Seriously();
-			source = seriously.source(sourcePixels, {
-				width: colors.length,
-				height: 2
-			});
-
-			canvas = document.createElement('canvas');
-			canvas.width = 5;
-			canvas.height = 2;
-			target = seriously.target(canvas);
-
-			effect = seriously.effect('chroma');
-			ok(effect, 'Chroma effect successfully created');
-
-			effect.source = source;
-			target.source = effect;
-			effect.balance = 0;
-			effect.clipBlack = 0.1;
-			effect.clipWhite = 0.88;
-
-			// test with both green and blue screen colors
-			['rgb(51, 204, 77)', 'rgb(51, 77, 204)'].forEach(function (screen, row) {
-				var matchOffset = row * colors.length * 4,
-					noMatchOffset = (1 - row) * colors.length * 4,
-					color = row ? 'Blue' : 'Green';
-
-				effect.screen = screen;
-				target.render();
-
-				try {
-					pixels = target.readPixels(0, 0, colors.length, 2, pixels);
-				} catch (e) {
-					error = e;
-				}
-				ok(incompatible ? error : !error, 'readPixels throws error iff incompatible');
-				if (!incompatible) {
-					equal(pixels[matchOffset + 3], 247, color + ' - Low saturation partially keyed out');
-					equal(pixels[matchOffset + 7], 255, color + ' - Very low saturation remains opaque');
-					equal(pixels[matchOffset + 11], 255, color + ' - Low saturation and max channel remains opaque');
-					equal(pixels[matchOffset + 15], 0, color + ' - High saturation fully keyed out');
-					equal(pixels[matchOffset + 19], 11, color + ' - Mid saturation partially keyed out');
-
-					equal(pixels[noMatchOffset + 3], 255, color + ' - Low saturation, different color remains opaque');
-					equal(pixels[noMatchOffset + 7], 255, color + ' - Very low saturation, different color remains opaque');
-					equal(pixels[noMatchOffset + 11], 255, color + ' - Low saturation and max channel, different color remains opaque');
-					equal(pixels[noMatchOffset + 15], 255, color + ' - High saturation, different color remains opaque');
-					equal(pixels[noMatchOffset + 19], 255, color + ' - Mid saturation, different color remains opaque');
-				}
-			});
-
-
-			seriously.destroy();
-			Seriously.removePlugin('chroma');
-
-			start();
+		seriously = new Seriously();
+		arraySource = seriously.source(array, {
+			width: 2,
+			height: 2
 		});
-	});
 
-	asyncTest('Channels', 5, function () {
-		require([
-			'seriously',
-			'effects/seriously.channels',
-			'src/sources/array'
-		], function (Seriously) {
-			var seriously,
-				effect,
-				target,
-				canvas,
-				pixels,
-				error,
-				incompatible,
-				array = [
-					255, 255, 255, 255,
-					255, 0, 0, 255,
-					0, 255, 0, 128,
-					0, 0, 255, 0
-				],
-				arraySource,
-				img,
-				imgSource,
-				expected;
+		canvas = document.createElement('canvas');
+		canvas.width = canvas.height = 1;
+		target = seriously.target(canvas);
 
-			incompatible = Seriously.incompatible();
+		img = document.createElement('img');
+		img.src = 'media/tiny.png';
+		imgSource = seriously.source(img);
 
-			seriously = new Seriously();
-			arraySource = seriously.source(array, {
-				width: 2,
-				height: 2
-			});
+		effect = seriously.effect('channels');
+		ok(effect, 'Channels effect successfully created');
 
-			canvas = document.createElement('canvas');
-			canvas.width = canvas.height = 1;
-			target = seriously.target(canvas);
+		effect.alphaSource = arraySource;
+		ok(true, 'Set alphaSource without having a main "source" input set (issue #109)');
 
-			img = document.createElement('img');
-			img.src = 'media/tiny.png';
-			imgSource = seriously.source(img);
+		effect.source = imgSource;
+		equal(effect.redSource, imgSource, 'Unset channel sources fall back to main "source" input');
 
-			effect = seriously.effect('channels');
-			ok(effect, 'Channels effect successfully created');
+		effect.source = arraySource;
+		equal(effect.source, arraySource, 'Set main "source" a second time (issue #109)');
+		equal(effect.redSource, imgSource, 'Unset channel sources remain set to original main "source" even after it changes.');
 
-			effect.alphaSource = arraySource;
-			ok(true, 'Set alphaSource without having a main "source" input set (issue #109)');
+		target.source = effect;
 
-			effect.source = imgSource;
-			equal(effect.redSource, imgSource, 'Unset channel sources fall back to main "source" input');
+		//todo: render and readPixels to test accurate render results
 
-			effect.source = arraySource;
-			equal(effect.source, arraySource, 'Set main "source" a second time (issue #109)');
-			equal(effect.redSource, imgSource, 'Unset channel sources remain set to original main "source" even after it changes.');
-
-			target.source = effect;
-
-			//todo: render and readPixels to test accurate render results
-
-			seriously.destroy();
-			Seriously.removePlugin('channels');
-
-			start();
-		});
+		seriously.destroy();
+		Seriously.removePlugin('channels');
 	});
 }());
